kali_agent_v4/requirements.txt
# KaliAgent v4 - Core Dependencies
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
requests>=2.31.0
reportlab>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
tqdm>=4.66.0

# Security Tools Integration
python-nmap>=0.7.1
scapy>=2.5.0
cryptography>=41.0.0

# Database
sqlalchemy>=2.0.0
redis>=5.0.0

# Async support
gevent>=23.9.0
gevent-websocket>=0.10.1

# Optional: AI/ML (uncomment if using advanced features)
# torch>=2.0.0
# transformers>=4.35.0

# Optional: Cloud integrations
# boto3>=1.34.0  # AWS
# azure-identity>=1.15.0  # Azure
# google-cloud>=0.34.0  # GCP
